project2/pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "scalar-grid-env"
version = "0.1.0"
description = "OpenEnv-compatible grid reinforcement learning environment with rewards, obstacles, and grading."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "openenv-core>=0.2.0",
  "pydantic>=2.10.6",
  "uvicorn>=0.34.0",
]

[project.scripts]
server = "server.app:main"

[tool.hatch.build.targets.wheel]
packages = ["scalar_grid_env", "server"]
include = [
  "__init__.py",
  "client.py",
  "models.py",
  "openenv.yaml",
  "README.md",
]

